Fix Chrome fullscreen error (#113)
When embedding h5p and using the fullscreen enter/exit feature, Chrome can produce the following error:
TypeError: 'caller', 'callee', and 'arguments' properties may not be accessed on strict mode functions or the arguments objects for calls to them
    at HTMLDocument.<anonymous> 

This happens because arguments.callee is not accessible. To fix this, it is sufficient to name the anonymous callback function and reference it by name.