Mixed version dependency conflict
When an H5P has dependencies on multiple versions of the same library,
the export will only generate with the first dependency it comes across,
rather than all versions.

This is because the keys used to identify which dependencies are needed
only relies on machine name, but not version number.