Fix resizing issue when flickering
This issue was first noticed in https://tracker.moodle.org/browse/MDL-84032
The root of the problem seems that when content is also responsive (font for example)
the resizing in some case creates an infinite loop that leads to constant
resizing of the activity leading to a flickering effect.
By using a threshold under which the iframe is not resized, we avoid the flickering.